move slack communication to server
