+++ .env.example
@@ -1,2 +1,3 @@
 ALPACA_API_KEY_ID = ""
 ALPACA_API_KEY_VALUE = ""
+SLACK_WEBHOOK_API_KEY = ""
